Closes #24
Added a few missing
