+++ .gitignore
@@ -263,3 +263,12 @@ MidasCivil_Engine/obj/Debug/MidasCivil_Engine.dll
 MidasCivil_Engine/obj/Debug/MidasCivil_Engine.pdb
 MidasCivil_Adapter/obj/Debug/MidasCivil_Adapter.dll
 MidasCivil_Adapter/obj/Debug/MidasCivil_Adapter.pdb
+.vs/MidasCivil_Toolkit/v14/.suo
+Build/MidasCivil_Adapter.dll
+Build/MidasCivil_Adapter.pdb
+Build/MidasCivil_Engine.dll
+Build/MidasCivil_Engine.pdb
+MidasCivil_Adapter/obj/Debug/MidasCivil_Adapter.dll
+MidasCivil_Adapter/obj/Debug/MidasCivil_Adapter.pdb
+MidasCivil_Engine/obj/Debug/MidasCivil_Engine.dll
+MidasCivil_Engine/obj/Debug/MidasCivil_Engine.pdb
