--- MidasCivil_Adapter/CRUD/Create/Loads/ConcentratedLoad.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using BH.oM.Structure.Loads;
-
-namespace BH.Adapter.MidasCivil
-{
-    public partial class MidasCivilAdapter
-    {
-        public string CreateConcentratedLoad(PointForce pointForce)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
