--- MidasCivil_Adapter/CRUD/Create/Loads/SelfWeight.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using BH.oM.Structure.Loads;
-
-namespace BH.Adapter.MidasCivil
-{
-    public partial class MidasCivilAdapter
-    {
-        public string CreateSelfWeight(GravityLoad gravityLoad)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
