+++ MidasCivil_Adapter/CRUD/Create/Properties/Materials.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BH.Adapter.MidasCivil.CRUD.Create.Properties
+{
+    class Materials
+    {
+    }
+}
