+++ MidasCivil_Adapter/MidasCivil_Adapter.csproj
@@ -71,8 +71,9 @@
   <ItemGroup>
     <Compile Include="CRUD\Create\Elements\Element.cs" />
     <Compile Include="CRUD\Create\Elements\Node.cs" />
-    <Compile Include="CRUD\Create\Loads\ConcentratedLoad.cs" />
-    <Compile Include="CRUD\Create\Loads\SelfWeight.cs" />
+    <Compile Include="CRUD\Create\Properties\Materials.cs" />
+    <Compile Include="CRUD\Create\Properties\SectionProperties.cs" />
+    <Compile Include="CRUD\Create\Properties\Supports.cs" />
     <Compile Include="CRUD\PrivateHelpers\CleanString.cs" />
     <Compile Include="CRUD\PrivateHelpers\GetSectionText.cs" />
     <Compile Include="CRUD\Create\_Create.cs" />
@@ -101,7 +102,7 @@
     <Compile Include="Types\DependencyTypes.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="CRUD\Create\Properties\" />
+    <Folder Include="CRUD\Create\Loads\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
